containerized registration and added comments
