+++ src/pages/Access.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react"
 import { Typography, Box } from "@mui/material"
-import Register from "../components/auth/register"
+import RegisterContainer from "../components/auth/Registration/RegisterContainer"
 import LoginContainer from "../components/auth/Login/LoginContainer"
 import vibeBuddiesLogo from "../components/auth/vibebuddies.png"
 import { useNavigate } from "react-router-dom" // Import useNavigate from react-router-dom
@@ -106,7 +106,7 @@ const Access: React.FC = () => {
         </Box>
 
         {/* conditionally renders Register or Login components based on which button is clicked */}
-        {showLogin ? <LoginContainer /> : <Register />}
+        {showLogin ? <LoginContainer /> : <RegisterContainer />}
       </Box>
     </Box>
   )
